requirements.txt
django==4.2.11
mysqlclient
python-decouple
django-bootstrap-v5
psycopg2-binary
Pillow